Refactored import overlay into a component
